requirements.txt
asgiref==3.3.4
Django==3.2.4
django-appconf==1.0.4
django-crispy-forms==1.12.0
django-select2==7.7.1
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
